# specs/001-tandem-slack-bot/contracts/api.yaml
openapi: 3.0.3
info:
  title: Tandem Slack Bot API
  description: API for managing tasks detected from Slack and scheduling them in Google Calendar
  version: 1.0.0
  contact:
    name: Tandem Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api-dev.tandem.app/api
    description: Development environment
  - url: https://api.tandem.app/api
    description: Production environment
security:
  - BearerAuth: []
paths:
...
  /auth/slack:
    get:
      tags: [Authentication]
      summary: Initiate Slack OAuth flow
      description: Redirects user to Slack OAuth authorization
      responses:
        '302':
          description: Redirect to Slack OAuth
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /auth/google:
    get:
      tags: [Authentication]
      summary: Initiate Google OAuth flow
      description: Redirects user to Google OAuth authorization
      responses:
        '302':
          description: Redirect to Google OAuth
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
          required: [code, message]
      required: [error]
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...